fix(og): server-render OG meta tags for X/Twitter crawler
The agent profile page was a client component so OG tags were set via
useEffect — invisible to crawlers. Extract interactive UI into
AgentProfile.tsx client component, make page.tsx a server component
with generateMetadata that fetches agent data from KV.
